pyproject.toml
[tool.poetry]
name = "qwery-core"
version = "0.1.0"
description = "Qwery Core – natural language to SQL agent with integrated visualization"
authors = ["Guepard"]
readme = "README.md"
packages = [{ include = "qwery_core", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.1"
plotly = "^5.22.0"
psycopg2-binary = "^2.9.9"
openai = "^2.7.1"
supabase = "^2.4.0"
anthropic = "^0.36.0"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
ruff = "^0.5.5"
mypy = "^1.10.0"
types-requests = "^2.32.0.20240712"

[tool.poetry.scripts]
qwery-cli = "qwery_core.cli:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

